requirements.txt
gdown==5.2.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.2
PyYAML==6.0.2
tqdm==4.67.1
imagehash==4.3.1
requests==2.32.3
beautifulsoup4==4.12.3
PySide6>=6.6.0
psd-tools>=1.9.0

